workspace/requirements/ingestion.txt
requests
kafka-python
minio
pytrends
pandas